Kernel: Merge Updates
nf: IDLETIMER: time-stamp and suspend/resume handling.

Message notifications contains an additional timestamp field in nano
seconds.
The expiry time for the timers are modified during suspend/resume.
If timer was supposed to expire while the system is suspended then
a notification is sent when it resumes with the timestamp of the
scheduled expiry.

Removes the race condition for multiple work scheduled.

Change-Id: I52fdd2800aaf88e477df1228474b4965a9397450
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Bug: 13247811

nf: time keeping changed from wall-clock to boot-time

Time stamps for the last modified and delay-triggered time store
boot-time instead of wall-clock time.

Change-Id: I256ae28dafac60add3b0b924c706ecdb71f5bbbf
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>

nf: Fix PM notifier list corruption

unregister the pm_notifier when the netfilter is destroyed.
check for null timer when system suspends.

Bug: 13508802
Bug: 13508117

Change-Id: I8ff0c18993d8374103fae6f35183c3eae8a7e6f0
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>

Revert "nf: Fix PM notifier list corruption"

This reverts commit 5e647ffb460d9301f05281ba0a5ca6e257129c8d.

Change-Id: Ied07d55456f813c54c6ed1a171329bd53e939f38

Revert "nf: IDLETIMER: time-stamp and suspend/resume handling."

This reverts commit b1027c0228924d23cc3918d43fbf9fbc21f0dbe9.

Conflicts:
	net/netfilter/xt_IDLETIMER.c

Change-Id: Ie98e5f1fcf1ef3c087021d3e24396f36457b2270

Remove build & conflict errors for commit 1d8c783dff935599870485f76fdfc76ee84799e9

Signed-off-by: Ruchi Kandoi<kandoiruchi@google.com>
Change-Id: I549228fa5b1b95d3fc2b4c62c264126923424811

nf: IDLETIMER: time-stamp and suspend/resume handling.

Message notifications contains an additional timestamp field in nano seconds.
The expiry time for the timers are modified during suspend/resume.
If timer was supposed to expire while the system is suspended then a
notification is sent when it resumes with the timestamp of the scheduled expiry.

Removes the race condition for multiple work scheduled.

Bug: 13247811

Change-Id: I752c5b00225fe7085482819f975cc0eb5af89bff
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>

nf: IDLETIMER: Adds the uid field in the msg

Message notifications contains an additional uid field. This field
represents the uid that was responsible for waking the radio. And hence
it is present only in notifications stating that the radio is now
active.

Change-Id: I18fc73eada512e370d7ab24fc9f890845037b729
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Bug: 20264396